Atualizada a API de Queries

+++ HTTP/BooksAPI/app.py
@@ -14,13 +14,12 @@ if testing:
 else:
     app.config['SQLALCHEMY_DATABASE_URI'] = os.getenv('DATABASE_URL')
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-app.config['SQLALCHEMY_COMMIT_ON_TEARDOWN'] = False
+app.json.ensure_ascii = False
 
 db = SQLAlchemy(app)
 ma = Marshmallow(app)
 migrate = Migrate()
 migrate.init_app(app, db)
-
 app.app_context().push()
 
 class Author(db.Model):
@@ -66,17 +65,17 @@ def add_author():
 
 @app.route('/authors', methods=['GET'])
 def get_authors():
-    authors = Author.query.all()
+    authors = db.session.execute(db.select(Author)).scalars()
     return authors_schema.jsonify(authors)
 
 @app.route('/authors/<int:author_id>', methods=['GET'])
 def get_author(author_id):
-    author = Author.query.get_or_404(author_id)
+    author = db.get_or_404(Author, author_id)
     return author_schema.jsonify(author)
 
 @app.route('/authors/<int:author_id>', methods=['PUT'])
 def update_author(author_id):
-    author = Author.query.get_or_404(author_id)
+    author = db.get_or_404(Author, author_id)
     data = request.get_json()
     author.name = data["name"]
     db.session.commit()
@@ -84,41 +83,44 @@ def update_author(author_id):
 
 @app.route('/authors/<int:author_id>', methods=['DELETE'])
 def delete_author(author_id):
-    author = Author.query.get_or_404(author_id)
-    db.session.delete(author)
-    db.session.commit()
-    return author_schema.jsonify(author)
+    author = db.get_or_404(Author, author_id)
+    if author:
+        db.session.delete(author)
+        db.session.commit()
+        return "Author deleted", 200
+    else:
+        return "Author not found", 404
 
 @app.route('/books', methods=['POST'])
 def add_book():
     data = request.get_json()
-    author = Author.query.get_or_404(data["author_id"])
-    book = Book(title=data["title"], genre=data["genre"], author=author)
+    author = db.get_or_404(Author, data["author_id"])
+    book = Book(title=data["title"], genre=data["genre"], author_id=author.id)
     db.session.add(book)
     db.session.commit()
     return book_schema.jsonify(book)
 
 @app.route('/book', methods=['GET'])
 def get_books():
-    all_books = Book.query.all()
+    all_books = db.session.execute(db.select(Book)).scalars()
     result = books_schema.dump(all_books)
     return jsonify(result)
 
 @app.route('/book/<id>', methods=['GET'])
 def get_book(id):
-    book = Book.query.get_or_404(id)
+    book = db.get_or_404(Book, id)
     return book_schema.jsonify(book)
 
 @app.route('/book/author/<author_id>', methods=['GET'])
 def get_book_by_author_id(author_id):
-    author = Author.query.get_or_404(author_id)
-    books = Book.query.filter_by(author=author)
+    author = db.get_or_404(Author, author_id)
+    books = db.session.execute(db.select(Book).filter_by(author=author)).scalars()
     return jsonify([{'title':book.title,'genre':book.genre} for book in books])
 
 @app.route('/books/<int:book_id>', methods=['PUT'])
 def update_book(book_id):
     data = request.get_json()
-    book = Book.query.get_or_404(book_id)
+    book = db.get_or_404(Book, book_id)
     if "title" in data:
         book.title = data["title"]
     if "author_id" in data:
@@ -131,12 +133,11 @@ def update_book(book_id):
 
 @app.route('/books/<int:book_id>', methods=['DELETE'])
 def delete_book(book_id):
-    book = Book.query.get_or_404(book_id)
+    book = db.get_or_404(Book, book_id)
     if book:
-        db.session.delete(book.author)
         db.session.delete(book)
         db.session.commit()
-        return book_schema.jsonify(book)
+        return "Book deleted", 200
     else:
         return "Book not found", 404
